Move keyword indexer for use with both profiles.
